Added XML serialization for Clovek class and list
The `System.Xml.Serialization` namespace was added to support XML serialization. The `Clovek` class was made serializable and a parameterless constructor was added for this purpose. The `telefonniSeznam` list instantiation was updated to use the new() syntax. Code was added to serialize and deserialize the `telefonniSeznam` list to and from an XML file, with proper disposal of `StreamWriter` and `StreamReader` after use.